requirements/development.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar
django-admin-tools
django-tinymce
# django-russian-admin
